refactor(background): separate translation and summary handlers
Keep queue creation, configuration watching, and batch execution in
translation-queues. Move webpage and subtitle request handlers into their
own modules and register the video-summary cache handlers independently.

Name the shared queued summary helper for its translation-context role.
Preserve existing message names, cache keys, queue behavior, and synchronous
handler registration during background startup.

Validation: 2859 tests passed, 4 skipped with SKIP_FREE_API=true;
type-aware lint/typecheck, formatting, and Chrome production build passed.
