+++ src/entrypoints/background/translation-queues.ts
@@ -2,52 +2,29 @@ import type { HostedAiTextStreamRoute } from "@/types/background-stream"
 import type { Config } from "@/types/config/config"
 import type { ProviderConfig } from "@/types/config/provider"
 import type { BatchQueueConfig, RequestQueueConfig } from "@/types/config/translate"
-import type { SubtitlePromptContext, WebPagePromptContext } from "@/types/content"
+import type { WebPagePromptContext } from "@/types/content"
 import type { PromptResolver } from "@/utils/host/translate/api/ai"
-import type { PromptableProviderRef, SerializableProviderRef } from "@/utils/providers/provider-ref"
+import type { SerializableProviderRef } from "@/utils/providers/provider-ref"
 import { LANG_CODE_TO_EN_NAME } from "@read-frog/definitions"
-import { browser, storage } from "#imports"
+import { storage } from "#imports"
 import { isLLMProviderConfig } from "@/types/config/provider"
 import { putBatchRequestRecord } from "@/utils/batch-request-record"
 import { CONFIG_STORAGE_KEY, DEFAULT_CONFIG } from "@/utils/constants/config"
-import {
-  BATCH_SEPARATOR,
-  BATCH_SEPARATOR_LINE_PATTERN,
-  isNoTranslationSentinel,
-} from "@/utils/constants/prompt"
+import { BATCH_SEPARATOR, BATCH_SEPARATOR_LINE_PATTERN } from "@/utils/constants/prompt"
 import {
   BATCH_TIMEOUT_BASE_MS,
   BATCH_TIMEOUT_PER_CHAR_MS,
   MAX_BATCH_TIMEOUT_MS,
 } from "@/utils/constants/translate"
-import { generateArticleSummary } from "@/utils/content/summary"
-import { cleanText } from "@/utils/content/utils"
 import { getRandomUUID } from "@/utils/crypto-polyfill"
-import { db } from "@/utils/db/dexie/db"
 import { Sha256Hex } from "@/utils/hash"
 import { executeTranslate } from "@/utils/host/translate/execute-translate"
-import {
-  assertHtmlAttributeMarkerIntegrity,
-  hasHtmlAttributeMarkerProtocol,
-  isHtmlAttributeMarkerIntegrityError,
-} from "@/utils/host/translate/html-attribute-markers"
-import {
-  auditInlineAtomTokens,
-  hasInlineAtomTokens,
-} from "@/utils/host/translate/inline-atom-tokens"
-import { normalizePromptContextValue } from "@/utils/host/translate/translate-text"
 import { logger } from "@/utils/logger"
-import { onMessage } from "@/utils/message"
 import { getSubtitlesTranslatePrompt } from "@/utils/prompts/subtitles"
 import { getTranslatePrompt } from "@/utils/prompts/translate"
-import {
-  canProviderRefGenerateText,
-  getProviderCacheIdentity,
-} from "@/utils/providers/provider-ref"
 import { BatchQueue } from "@/utils/request/batch-queue"
-import { CancelledScopeRegistry, TranslationCancelledError } from "@/utils/request/cancellation"
+import { CancelledScopeRegistry } from "@/utils/request/cancellation"
 import { RequestQueue } from "@/utils/request/request-queue"
-import { generateTextForProviderRef } from "./background-stream"
 import { runStreamTextInBackground } from "./background-stream"
 import { ensureInitializedConfig } from "./config"
 
@@ -59,7 +36,7 @@ function isSerializedPageProvider(
   return "kind" in provider && (provider.kind === "local" || provider.kind === "system")
 }
 
-function getLocalProviderConfig(provider: QueuedTranslationProvider): ProviderConfig | null {
+export function getLocalProviderConfig(provider: QueuedTranslationProvider): ProviderConfig | null {
   if (!isSerializedPageProvider(provider)) return provider
   return provider.kind === "local" ? provider.config : null
 }
@@ -125,27 +102,6 @@ export function shouldUseBatchQueue(provider: QueuedTranslationProvider): boolea
   return local ? isLLMProviderConfig(local) : true
 }
 
-async function getValidatedCachedTranslation(
-  hash: string,
-  sourceText: string,
-  validateHtmlAttributeMarkers: boolean,
-): Promise<string | undefined> {
-  const cached = await db.translationCache.get(hash)
-  if (!cached) return undefined
-  if (!validateHtmlAttributeMarkers) return cached.translation
-
-  try {
-    assertHtmlAttributeMarkerIntegrity(sourceText, cached.translation)
-    return cached.translation
-  } catch (error) {
-    if (!isHtmlAttributeMarkerIntegrityError(error)) throw error
-
-    await db.translationCache.delete(hash)
-    logger.warn("Deleted cached translation with invalid HTML attribute markers", error)
-    return undefined
-  }
-}
-
 export async function executeBatchTranslation<TContext>(
   dataList: TranslateBatchData<TContext>[],
   promptResolver: PromptResolver<TContext>,
@@ -165,95 +121,6 @@ export async function executeBatchTranslation<TContext>(
     hostedFeature,
   })
   return parseBatchResult(result)
-}
-
-/**
- * One cached, queued summary generation for both the page summary and the
- * video summary. The two were byte-for-byte identical apart from which parts
- * went into the cache key, so the key parts are the parameter and everything
- * else — the double-check inside the thunk, the empty-string sentinel that
- * keeps a failed generation from being cached as a hit, the queue admission —
- * is shared.
- */
-async function getOrGenerateSummary(args: {
-  title: string
-  textContent: string
-  providerRef: PromptableProviderRef
-  hostedFeature: HostedAiTextStreamRoute
-  cacheKeyParts: string[]
-  requestQueue: RequestQueue
-}): Promise<string | null> {
-  const { title, textContent, providerRef, hostedFeature, cacheKeyParts, requestQueue } = args
-  const preparedText = cleanText(textContent)
-  if (!preparedText) {
-    return null
-  }
-
-  const cacheKey = Sha256Hex(...cacheKeyParts, getProviderCacheIdentity(providerRef))
-
-  const cached = await db.articleSummaryCache.get(cacheKey)
-  if (cached) {
-    logger.info("Using cached summary")
-    return cached.summary
-  }
-
-  // Stable for this queue task, mirroring the translation batches: automatic
-  // RequestQueue retries must reuse the idempotency key because the first
-  // hosted response may have been lost after billing.
-  const hostedRequestId = providerRef.kind === "system" ? getRandomUUID() : undefined
-
-  const thunk = async (signal?: AbortSignal) => {
-    const cachedAgain = await db.articleSummaryCache.get(cacheKey)
-    if (cachedAgain) {
-      return cachedAgain.summary
-    }
-
-    const summary = await generateArticleSummary(title, textContent, providerRef, {
-      hostedFeature,
-      signal,
-      generate: (payload, runOptions) =>
-        generateTextForProviderRef({ ...payload, requestId: hostedRequestId }, runOptions),
-    })
-    if (!summary) {
-      return ""
-    }
-
-    await db.articleSummaryCache.put({
-      key: cacheKey,
-      summary,
-      createdAt: new Date(),
-    })
-
-    logger.info("Generated and cached new summary")
-    return summary
-  }
-
-  try {
-    const summary = await requestQueue.enqueue(thunk, Date.now(), cacheKey)
-    return summary || null
-  } catch (error) {
-    logger.warn("Failed to get/generate summary:", error)
-    return null
-  }
-}
-
-/**
- * Bumped whenever the summary prompt changes: the cache key has to move with
- * it or every user keeps being served answers from the old wording.
- */
-const VIDEO_SUMMARY_PROMPT_VERSION = "1"
-
-function videoSummaryCacheKey(args: {
-  transcript: string
-  targetLanguage: string
-  providerRef: PromptableProviderRef
-}): string {
-  return Sha256Hex(
-    Sha256Hex(args.transcript),
-    args.targetLanguage,
-    VIDEO_SUMMARY_PROMPT_VERSION,
-    getProviderCacheIdentity(args.providerRef),
-  )
 }
 
 export interface TranslateBatchData<TContext = unknown> {
@@ -485,7 +352,7 @@ const selectWebPageQueueConfig = (config: Config) => ({
   batchQueueConfig: config.pageTranslation.batchQueueConfig,
 })
 
-export function setUpWebPageTranslationQueue(): void {
+export function createWebPageTranslationQueues() {
   // Scopes whose cancel already drained the queues. Consulted by (a) the
   // enqueue handler after its cache-lookup await and (b) the batch queue's
   // retry/fallback path after its backoff sleep — both are windows where a
@@ -493,13 +360,11 @@ export function setUpWebPageTranslationQueue(): void {
   // would otherwise be lost (#1881).
   const cancelledScopes = new CancelledScopeRegistry()
 
-  type WebTranslationPromptContext = WebPagePromptContext
-
-  const webPromptResolver: PromptResolver<WebTranslationPromptContext> = getTranslatePrompt
+  const webPromptResolver: PromptResolver<WebPagePromptContext> = getTranslatePrompt
 
   const queuesPromise = loadQueueSetupConfig("webpage", selectWebPageQueueConfig).then(
     ({ requestQueueConfig, batchQueueConfig, configSource }) =>
-      createTranslationQueues<WebTranslationPromptContext>({
+      createTranslationQueues<WebPagePromptContext>({
         requestQueueConfig,
         batchQueueConfig,
         promptResolver: webPromptResolver,
@@ -510,183 +375,9 @@ export function setUpWebPageTranslationQueue(): void {
       }),
   )
 
-  // Everything below registers in the FIRST synchronous turn of the SW: an
-  // MV3 wake-triggering message can no longer be dropped while init awaits.
   watchQueueConfig("webpage", queuesPromise, selectWebPageQueueConfig)
 
-  onMessage("enqueueTranslateRequest", async (message) => {
-    const { requestQueue, batchQueue } = await queuesPromise
-    const {
-      data: {
-        text,
-        langConfig,
-        providerRef,
-        scheduleAt,
-        hash,
-        textFormat,
-        preserveLineBreaks,
-        webTitle,
-        webDescription,
-        webContent,
-        webSummary,
-        sessionId,
-        forceRetranslation = false,
-        hostedFeature,
-      },
-    } = message
-    const scope = buildTranslationScopeKey(message.sender, sessionId)
-
-    const validateHtmlAttributeMarkers =
-      textFormat === "html" && hasHtmlAttributeMarkerProtocol(text)
-    if (validateHtmlAttributeMarkers) {
-      assertHtmlAttributeMarkerIntegrity(text, text)
-    }
-
-    // Check cache first — unless the user asked for a fresh translation. The
-    // existing entry is left untouched unless the fresh request succeeds below.
-    if (hash && !forceRetranslation) {
-      const cachedTranslation = await getValidatedCachedTranslation(
-        hash,
-        text,
-        validateHtmlAttributeMarkers,
-      )
-      if (cachedTranslation !== undefined) return cachedTranslation
-    }
-
-    // The cache lookup above yielded — the session's cancel may have drained
-    // the queues while this handler was suspended. Enqueueing now would park
-    // an undraininable task on a dead scope, so abort instead (the content
-    // side swallows this error).
-    if (scope && cancelledScopes.has(scope)) {
-      throw new TranslationCancelledError(scope)
-    }
-
-    let result: string
-    const context: WebTranslationPromptContext = {
-      webTitle: normalizePromptContextValue(webTitle),
-      webDescription: normalizePromptContextValue(webDescription),
-      webContent: normalizePromptContextValue(webContent),
-      webSummary: normalizePromptContextValue(webSummary),
-    }
-
-    if (shouldUseBatchQueue(providerRef)) {
-      const data = {
-        text,
-        langConfig,
-        provider: providerRef,
-        hash,
-        scheduleAt,
-        context,
-        scope,
-        hostedFeature,
-      }
-      result = await batchQueue.enqueue(data)
-    } else {
-      const localProvider = getLocalProviderConfig(providerRef)
-      if (!localProvider) {
-        throw new Error("Invalid local page translation provider")
-      }
-      // Create thunk based on type and params
-      const thunk = (signal?: AbortSignal) =>
-        executeTranslate(text, langConfig, localProvider, getTranslatePrompt, {
-          textFormat,
-          preserveLineBreaks,
-          signal,
-        })
-      result = await requestQueue.enqueue(thunk, scheduleAt, hash, scope ? [scope] : undefined)
-    }
-
-    if (validateHtmlAttributeMarkers) {
-      assertHtmlAttributeMarkerIntegrity(text, result)
-    }
-
-    // A response that dropped, invented or duplicated an inline-atom
-    // placeholder is still rendered by the content script (the formulas are
-    // appended) but must not be persisted: the next visit deserves a fresh
-    // attempt instead of a permanently degraded paragraph. The sentinel is
-    // exempt — "no translation needed" carries no placeholders by definition,
-    // and auditing it as a total loss would make every already-in-target-language
-    // paragraph containing a formula re-hit the provider on every page load.
-    const inlineAtomTokensIntact =
-      !hasInlineAtomTokens(text) ||
-      isNoTranslationSentinel(result) ||
-      auditInlineAtomTokens(text, result).ok
-    if (result && !inlineAtomTokensIntact) {
-      logger.warn("Inline atom placeholders were not preserved; result not cached")
-    }
-
-    // Cache the translation result if successful
-    if (result && hash && inlineAtomTokensIntact) {
-      await db.translationCache.put({
-        key: hash,
-        translation: result,
-        createdAt: new Date(),
-      })
-    }
-
-    return result
-  })
-
-  onMessage("getOrGenerateWebPageSummary", async (message) => {
-    const { requestQueue } = await queuesPromise
-    const { webTitle, webContent, providerRef } = message.data
-
-    if (!webTitle || !webContent) {
-      return null
-    }
-
-    // The payload type forces senders to narrow, but the wire is a trust
-    // boundary (a pre-update content script can send a translate-only ref).
-    // Refuse like the subtitle summary handlers do: a queue task for a
-    // provider with no model to prompt can only ever throw, after burning
-    // its retries.
-    if (!canProviderRefGenerateText(providerRef)) {
-      return null
-    }
-
-    return await getOrGenerateSummary({
-      title: webTitle,
-      textContent: webContent,
-      providerRef,
-      // The summary bills against the feature that triggered it (it is a
-      // sub-call of that feature, not a feature of its own); the sender names
-      // that route so the gate it serialized the ref under and the billing here
-      // cannot diverge. Absent only from a pre-update content script — page
-      // translation is the historical biller.
-      hostedFeature: message.data.hostedFeature ?? "pageTranslation",
-      cacheKeyParts: [webTitle, Sha256Hex(cleanText(webContent))],
-      requestQueue,
-    })
-  })
-
-  onMessage("cancelPageTranslationRequests", async (message) => {
-    const scope = buildTranslationScopeKey(message.sender, message.data.sessionId)
-    if (!scope) return
-    // Remember the scope BEFORE any await so enqueue handlers suspended on
-    // the cache lookup refuse to enqueue after this drain.
-    cancelledScopes.markScope(scope)
-    const { requestQueue, batchQueue } = await queuesPromise
-    // Batch queue first so pending batches cannot flush new request-queue
-    // tasks between the two drains.
-    const cancelledBatch = batchQueue.cancelByScope(scope)
-    const cancelledRequests = requestQueue.cancelByScope(scope)
-    if (cancelledBatch + cancelledRequests > 0) {
-      logger.info(
-        `Cancelled ${cancelledBatch + cancelledRequests} page-translation requests (scope: ${scope})`,
-      )
-    }
-  })
-
-  // A closed tab can never send its cancel message — sweep every scope the
-  // tab ever registered (#1881).
-  browser.tabs.onRemoved.addListener((tabId) => {
-    const prefix = `${tabId}:`
-    cancelledScopes.markPrefix(prefix)
-    void queuesPromise.then(({ requestQueue, batchQueue }) => {
-      batchQueue.cancelWhere((scope) => scope.startsWith(prefix))
-      requestQueue.cancelWhere((scope) => scope.startsWith(prefix))
-    })
-  })
+  return { queuesPromise, cancelledScopes }
 }
 
 const selectSubtitlesQueueConfig = (config: Config) => ({
@@ -695,9 +386,9 @@ const selectSubtitlesQueueConfig = (config: Config) => ({
 })
 
 /**
- * Set up subtitles translation queue and message handlers
+ * Create subtitle translation queues and watch their configuration
  */
-export function setUpSubtitlesTranslationQueue(): void {
+export function createSubtitlesTranslationQueues() {
   const queuesPromise = loadQueueSetupConfig("subtitles", selectSubtitlesQueueConfig).then(
     ({ requestQueueConfig, batchQueueConfig, configSource }) =>
       createTranslationQueues({
@@ -712,107 +403,5 @@ export function setUpSubtitlesTranslationQueue(): void {
 
   watchQueueConfig("subtitles", queuesPromise, selectSubtitlesQueueConfig)
 
-  onMessage("enqueueSubtitlesTranslateRequest", async (message) => {
-    const { requestQueue, batchQueue } = await queuesPromise
-    const {
-      data: { text, langConfig, providerRef, scheduleAt, hash, webTitle, webDescription, summary },
-    } = message
-
-    if (hash) {
-      const cached = await db.translationCache.get(hash)
-      if (cached) {
-        return cached.translation
-      }
-    }
-
-    let result: string
-    const context: SubtitlePromptContext = {
-      webTitle: normalizePromptContextValue(webTitle),
-      webDescription: normalizePromptContextValue(webDescription),
-      videoSummary: normalizePromptContextValue(summary),
-    }
-
-    if (shouldUseBatchQueue(providerRef)) {
-      const data = {
-        text,
-        langConfig,
-        provider: providerRef,
-        hash,
-        scheduleAt,
-        context,
-      }
-      result = await batchQueue.enqueue(data)
-    } else {
-      // Unreachable for system refs — shouldUseBatchQueue always batches them —
-      // but it must fail loudly rather than silently mistranslate if that ever
-      // changes, since executeTranslate only understands a local config.
-      const localConfig = getLocalProviderConfig(providerRef)
-      if (!localConfig) {
-        throw new Error("Built-in AI subtitle translation must use the batch queue")
-      }
-      const thunk = (signal?: AbortSignal) =>
-        executeTranslate(text, langConfig, localConfig, getSubtitlesTranslatePrompt, { signal })
-      result = await requestQueue.enqueue(thunk, scheduleAt, hash)
-    }
-
-    if (result && hash) {
-      await db.translationCache.put({
-        key: hash,
-        translation: result,
-        createdAt: new Date(),
-      })
-    }
-
-    return result
-  })
-
-  onMessage("getSubtitlesSummary", async (message) => {
-    const { requestQueue } = await queuesPromise
-    const { videoTitle, subtitlesContext, providerRef } = message.data
-
-    // Guarded on both sides: the content script should not send this for a
-    // provider with no model to prompt, and the queue must not admit a task
-    // that can only throw if it does.
-    if (!videoTitle || !subtitlesContext || !canProviderRefGenerateText(providerRef)) {
-      return null
-    }
-
-    return await getOrGenerateSummary({
-      title: videoTitle,
-      textContent: subtitlesContext,
-      providerRef,
-      hostedFeature: "videoSubtitles",
-      // Deliberately without the title, matching the previous key: a video's
-      // transcript identifies it, and including a title that players mutate
-      // would miss the cache on every re-render.
-      cacheKeyParts: [Sha256Hex(cleanText(subtitlesContext))],
-      requestQueue,
-    })
-  })
-
-  onMessage("getCachedVideoSummary", async (message) => {
-    const { transcript, targetLanguage, providerRef } = message.data
-    if (!transcript) {
-      return null
-    }
-
-    const cached = await db.articleSummaryCache.get(
-      videoSummaryCacheKey({ transcript, targetLanguage, providerRef }),
-    )
-    return cached?.summary ?? null
-  })
-
-  onMessage("saveVideoSummary", async (message) => {
-    const { transcript, targetLanguage, providerRef, summary } = message.data
-    const trimmed = summary.trim()
-    if (!transcript || !trimmed) {
-      return
-    }
-
-    await db.articleSummaryCache.put({
-      key: videoSummaryCacheKey({ transcript, targetLanguage, providerRef }),
-      summary: trimmed,
-      createdAt: new Date(),
-    })
-  })
+  return queuesPromise
 }
